move translation component functions to service
